+++ src/pyerasure/finite_field/binary.py
@@ -55,6 +55,22 @@ class Binary:
             elements[index // 8] |= 1 << (index % 8)
 
     @classmethod
+    def add(cls, lhs: int, rhs: int) -> int:
+        """Add two elements."""
+        if lhs < 0 or lhs > cls.max_value:
+            raise ValueError("lhs must be 0 and 1")
+        return lhs ^ rhs
+
+    @classmethod
+    def divide(cls, lhs: int, rhs: int) -> int:
+        """Divide two elements."""
+        if lhs < 0 or lhs > cls.max_value:
+            raise ValueError("lhs must be 0 or 1")
+        if rhs < 0 or rhs > cls.max_value:
+            raise ValueError("rhs must be 0 or 1")
+        return lhs
+
+    @classmethod
     def vector_add_into(cls, x: bytearray, y: bytearray):
         """Add y into x."""
         if len(x) != len(y):